pull out ColumnFamily
