+++ flow/src/main/scala/org/alephium/flow/io/Storages.scala
@@ -25,9 +25,8 @@ import org.alephium.io.IOResult
 import org.alephium.protocol.Hash
 import org.alephium.protocol.config.GroupConfig
 import org.alephium.protocol.vm.WorldState
-import org.alephium.storage.KeyValueSource
+import org.alephium.storage.{ColumnFamily, KeyValueSource}
 import org.alephium.storage.rocksdb.{RocksDBKeyValueStorage, RocksDBSource}
-import org.alephium.storage.rocksdb.RocksDBSource.ColumnFamily
 import org.alephium.util.AVector
 
 object Storages {
