+++ flow/src/test/scala/org/alephium/flow/io/ReadyTxStorageSpec.scala
@@ -20,8 +20,8 @@ import org.alephium.flow.model.ReadyTxInfo
 import org.alephium.flow.setting.AlephiumConfigFixture
 import org.alephium.protocol.{Generators, Hash}
 import org.alephium.protocol.model.ChainIndex
+import org.alephium.storage.ColumnFamily
 import org.alephium.storage.rocksdb.RocksDBSource
-import org.alephium.storage.rocksdb.RocksDBSource.ColumnFamily
 import org.alephium.util.{AlephiumSpec, AVector, TimeStamp}
 
 class ReadyTxStorageSpec
